+++ ChatdollKit/Scripts/IO/WakeWordListenerBase.cs
@@ -8,8 +8,7 @@ using ChatdollKit.Network;
 
 namespace ChatdollKit.IO
 {
-    [RequireComponent(typeof(VoiceRecorder))]
-    public class WakeWordListenerBase : MonoBehaviour
+    public class WakeWordListenerBase : VoiceRecorderBase
     {
         public List<string> WakeWords;
         public Func<string, Task> OnRecognizedAsync;
@@ -19,7 +18,6 @@ namespace ChatdollKit.IO
         public bool PrintResult = false;
 
         [Header("Voice Recorder Settings")]
-        public int SamplingFrequency = 16000;
         public float VoiceDetectionThreshold = 0.1f;
         public float VoiceDetectionMinimumLength = 0.2f;
         public float SilenceDurationToEndRecording = 0.3f;
@@ -32,64 +30,71 @@ namespace ChatdollKit.IO
         public Action<Exception> OnError = (e) => { Debug.LogError($"Recording wakeword error: {e.Message}\n{e.StackTrace}"); };
 
         // Private and protected members for recording voice and recognize task
-        private VoiceRecorder voiceRecorder;
         private CancellationTokenSource cancellationTokenSource;
         protected ChatdollHttp client = new ChatdollHttp();
 
-
-        private void Awake()
+        protected override void OnDestroy()
         {
-            voiceRecorder = gameObject.GetComponent<VoiceRecorder>();
-        }
-
-        private void OnDestroy()
-        {
+            base.OnDestroy();
             cancellationTokenSource?.Cancel();
         }
 
         public async Task StartListeningAsync()
         {
+            StartListening();   // Start recorder here to asure that GetVoiceAsync will be called after recorder started
+
             if (OnWakeAsync == null)
             {
                 Debug.LogError("OnWakeAsync must be set");
                 return;
             }
 
-            var voiceRecorderRequest = new VoiceRecorderRequest()
-            {
-                SamplingFrequency = SamplingFrequency,
-                VoiceDetectionThreshold = VoiceDetectionThreshold,
-                VoiceDetectionMinimumLength = VoiceDetectionMinimumLength,
-                SilenceDurationToEndRecording = SilenceDurationToEndRecording,
-                OnListeningStart = OnListeningStart,
-                OnListeningStop = OnListeningStop,
-                OnRecordingStart = OnRecordingStart,
-                OnDetectVoice = OnDetectVoice,
-                OnRecordingEnd = OnRecordingEnd,
-                OnError = OnError,
-            };
-
             cancellationTokenSource = new CancellationTokenSource();
             var token = cancellationTokenSource.Token;
 
             while (!token.IsCancellationRequested)
             {
-                var voiceRecorderResponse = await voiceRecorder.GetVoiceAsync(voiceRecorderRequest, token);
+                voiceDetectionThreshold = VoiceDetectionThreshold;
+                voiceDetectionMinimumLength = VoiceDetectionMinimumLength;
+                silenceDurationToEndRecording = SilenceDurationToEndRecording;
+                onListeningStart = OnListeningStart;
+                onListeningStop = OnListeningStop;
+                onRecordingStart = OnRecordingStart;
+                onDetectVoice = OnDetectVoice;
+                onRecordingEnd = OnRecordingEnd;
+                onError = OnError;
+
+                var voiceRecorderResponse = await GetVoiceAsync(0.0f, token);
                 if (voiceRecorderResponse != null && voiceRecorderResponse.Voice != null)
                 {
-                    // Recognize speech
-                    var recognizedText = await RecognizeSpeechAsync(voiceRecorderResponse.Voice);
-                    await OnRecognizedAsync?.Invoke(recognizedText);
-                    if (PrintResult)
+                    _ = ProcessVoiceAsync(voiceRecorderResponse.Voice); // Do not await to continue listening
+                }
+            }
+        }
+
+        private async Task ProcessVoiceAsync(AudioClip voice)
+        {
+            // Recognize speech
+            var recognizedText = await RecognizeSpeechAsync(voice);
+            if (OnRecognizedAsync != null)
+            {
+                await OnRecognizedAsync(recognizedText);
+            }
+            if (PrintResult)
+            {
+                Debug.Log($"Recognized(WakeWordListener): {recognizedText}");
+            }
+            foreach (var ww in WakeWords)
+            {
+                if (recognizedText.Contains(ww))
+                {
+                    try
                     {
-                        Debug.Log($"Recognized(WakeWordListener): {recognizedText}");
+                        await OnWakeAsync();
                     }
-                    foreach (var ww in WakeWords)
+                    catch (Exception ex)
                     {
-                        if (recognizedText.Contains(ww))
-                        {
-                            await OnWakeAsync();
-                        }
+                        Debug.LogError($"OnWakeAsync failed: {ex.Message}\n{ex.StackTrace}");
                     }
                 }
             }
